+++ examples/matrices.js
@@ -38,7 +38,7 @@ print(e);                                   // [[5, 6], [1, 1]]
 
 // set a submatrix.
 // Matrix indexes are zero-based.
-e.subset(math.index(1, [0, 2]), [[7, 8]]);
+e.subset(math.index(1, [0, 1]), [[7, 8]]);
 print(e);                                   // [[5, 6], [7, 8]]
 var f = math.multiply(d, e);
 print(f);                                   // [[19, 22], [43, 50]]
@@ -50,35 +50,39 @@ console.log();
 // Matrix indexes are zero-based.
 console.log('get a sub matrix');
 var h = math.diag(math.range(1,4));
-print(h);                                   // [[1, 0, 0], [0, 2, 0], [0, 0, 3]]
-print(h.subset(math.index([1, 3], [1, 3])));// [[2, 0], [0, 3]]
+print(h);                                     // [[1, 0, 0], [0, 2, 0], [0, 0, 3]]
+print(h.subset( math.index([1, 2], [1, 2]))); // [[2, 0], [0, 3]]
+var i = math.range(1,6);
+print(i);                                     // [1, 2, 3, 4, 5]
+print(i.subset(math.index(math.range(1,4)))); // [2, 3, 4]
 console.log();
+
 
 // resize a multi dimensional matrix
 console.log('resizing a matrix');
-var i = math.matrix();
+var j = math.matrix();
 var defaultValue = 0;
-i.resize([2, 2, 2], defaultValue);
-print(i);                                   // [[[0, 0], [0, 0]], [[0, 0], [0, 0]]]
-print(i.size());                            // [2, 2, 2]
-i.resize([2, 2]);
-print(i);                                   // [[0, 0], [0, 0]]
-print(i.size());                            // [2, 2]
+j.resize([2, 2, 2], defaultValue);
+print(j);                                   // [[[0, 0], [0, 0]], [[0, 0], [0, 0]]]
+print(j.size());                            // [2, 2, 2]
+j.resize([2, 2]);
+print(j);                                   // [[0, 0], [0, 0]]
+print(j.size());                            // [2, 2]
 console.log();
 
 // setting a value outside the matrices range will resize the matrix.
 // new elements will be initialized with zero.
 console.log('set a value outside a matrices range');
-var j = math.matrix();
-j.subset(math.index(2), 6);
-print(j);                                   // [0, 0, 6]
+var k = math.matrix();
+k.subset(math.index(2), 6);
+print(k);                                   // [0, 0, 6]
 console.log();
 
 console.log('set a value outside a matrices range, leaving new entries uninitialized');
-var k = math.matrix();
+var m = math.matrix();
 defaultValue = math.uninitialized;
-k.subset(math.index(2), 6, defaultValue);
-print(k);                                   // [undefined, undefined, 6]
+m.subset(math.index(2), 6, defaultValue);
+print(m);                                   // [undefined, undefined, 6]
 console.log();
 
 // create ranges