+++ test/function/arithmetic/multiply.test.js
@@ -132,8 +132,8 @@ describe('multiply', function() {
     });
 
     it('should multiply mixed fractions and numbers', function() {
-      assert.strictEqual(multiply(2, math.fraction(1,3)), 0.6666666666666666);
-      assert.strictEqual(multiply(math.fraction(1,3), 2), 0.6666666666666666);
+      assert.deepEqual(multiply(2, math.fraction(1,3)), math.fraction(2,3));
+      assert.deepEqual(multiply(math.fraction(1,3), 2), math.fraction(2,3));
     });
 
     it('should multiply a number and a unit correctly', function() {
@@ -154,6 +154,22 @@ describe('multiply', function() {
       assert.equal(multiply(unit('inch'), 2).toString(), '2 inch');
     });
 
+    it('should multiply two units correctly', function() {
+      assert.equal(multiply(unit('2 m'), unit('4 m')).toString(), '8 m^2');
+      assert.equal(multiply(unit('2 ft'), unit('4 ft')).toString(), '8 ft^2');
+      assert.equal(multiply(unit('65 mi/h'), unit('2 h')).to('mi').toString(), '130 mi');
+      assert.equal(multiply(unit('2 L'), unit('1 s^-1')).toString(), '2 L / s');
+      assert.equal(multiply(unit('2 m/s'), unit('0.5 s/m')).toString(), '1');
+    });
+
+    it('should multiply valueless units correctly', function() {
+      assert.equal(multiply(unit('m'), unit('4 m')).toString(), '4 m^2');
+      assert.equal(multiply(unit('ft'), unit('4 ft')).format(5), '4 ft^2');
+      assert.equal(multiply(unit('65 mi/h'), unit('h')).to('mi').toString(), '65 mi');
+      assert.equal(multiply(unit('2 L'), unit('s^-1')).toString(), '2 L / s');
+      assert.equal(multiply(unit('m/s'), unit('h/m')).toString(), '(m h) / (s m)');
+    });
+
     // TODO: cleanup once decided to not downgrade BigNumber to number
     it.skip('should multiply a bignumber and a unit correctly', function() {
       assert.equal(multiply(bignumber(2), unit('5 mm')).toString(), '10 mm');
@@ -172,11 +188,14 @@ describe('multiply', function() {
       assert.equal(multiply(unit('inch'), bignumber(2)).toString(), '2 inch');
     });
 
+
     it('should throw an error in case of unit non-numeric argument', function() {
-      assert.throws(function () {multiply(math.unit('5cm'), math.unit('4cm'));}, /TypeError: Unexpected type/);
+      // Multiplying two units is supported now --ericman314
+      //assert.throws(function () {multiply(math.unit('5cm'), math.unit('4cm'));}, /TypeError: Unexpected type/);
       assert.throws(function () {multiply(math.unit('5cm'), math.complex('2+3i'));}, /TypeError: Unexpected type/);
       assert.throws(function () {multiply(math.complex('2+3i'), math.unit('5cm'));}, /TypeError: Unexpected type/);
     });
+
 
     it('should throw an error if used with strings', function() {
       assert.throws(function () {multiply("hello", "world");});