+++ HISTORY.md
@@ -1,11 +1,26 @@
 # History
 
 
-## not yet released, version 2.0.0-SNAPSHOT
+## not yet released, version 2.1.0
 
-- String input is now converted to numbers by default for all functions. 
-- Adding two strings will no longer concatenate them, but will convert the 
-  strings to numbers and add them.
+- Implemented derived units (like `110 km/h in m/s`). Thanks @ericman314.
+- Implemented function `kldivergence` (Kullback-Leibler divergence). 
+  Thanks @saromanov.
+- Implemented function `mode`. Thanks @kv-kunalvyas.
+
+
+## 2015-07-29, version 2.0.1
+
+- Fixed operations with mixed fractions and numbers be converted to numbers
+  instead of fractions.
+
+
+## 2015-07-28, version 2.0.0
+
+- Large internal refactoring:
+  - performance improvements.
+  - allows to create custom bundles
+  - functions are composed using `typed-function` and are extensible
 - Implemented support for fractions, powered by the library `fraction.js`.
 - Implemented matrix LU decomposition with partial pivoting and a LU based 
   linear equations solver (functions `lup` and `lusolve`). Thanks @rjbaucells.
@@ -17,11 +32,18 @@
 - Implemented support for complex values to `nthRoot`. Thanks @gangachris.
 - Implemented util functions `isInteger`, `isNegative`, `isNumeric`, 
   `isPositive`, and `isZero`.
-- Large internal refactoring, allowing to create custom bundles of math.js.
+
+### breaking changes
+
+- String input is now converted to numbers by default for all functions. 
+- Adding two strings will no longer concatenate them, but will convert the 
+  strings to numbers and add them.
+- Function `index` does no longer accept an array `[start, end, step]`, but
+  instead accepts an array with arbitrary index values. It also accepts
+  a `Range` object as input.
 - Function `typeof` no longer returns lower case names, but now returns lower
   case names for primitives (like `number`, `boolean`, `string`), and 
   upper-camel-case for non-primitives (like `Array`, `Complex`, `Function`).
-- Function `import` now returns the imported objects.
 - Function `import` no longer supports a module name as argument. Instead,
   modules can be loaded using require: `math.import(require('module-name'))`.
 - Function `import` has a new option `silent` to ignore errors, and throws
@@ -31,8 +53,14 @@
 - Function `sum` now returns zero when input is an empty array. Thanks @FSMAxB.
 - The size of Arrays is no longer validated. Matrices will validate this on
   creation.
-- Fixed #397: Inaccuracies in `nthRoot` for very large values, and wrong results 
-  for very small values.
+
+
+## 2015-07-12, version 1.7.1
+
+- Fixed #397: Inaccuracies in nthRoot for very large values, and wrong results 
+  for very small values. (backported from v2)
+- Fixed #405: Parser throws error when defining a function in a multiline 
+  expression.
 
 
 ## 2015-05-31, version 1.7.0